refactor: split queries into individual files (user.js, file.js) and centralized Prisma client in prisma.js
