+++ src/db/queries/file.js
@@ -1,33 +1,4 @@
-import { PrismaClient } from '@prisma/client';
-const prisma = new PrismaClient();
-import { hashPassword, comparePassword } from '../utils/hash.js';
-
-async function getLoginUser(username, password) {
-    const user = await prisma.user.findUnique({
-        where: {
-            username: username,
-        }
-    });
-
-    if (user && await comparePassword(password, user.password)) {
-        return user;
-    } else {
-        return null;
-    }
-}
-
-async function postRegisterUser(username, name, email, password) {
-    const hashedPwd = await hashPassword(password);
-    const user = await prisma.user.create({
-        data: {
-            name: name,
-            username: username,
-            email: email,
-            password: hashedPwd,
-        }
-    })
-    return user;
-}
+import prisma from "./prisma.js";
 
 async function insertFile({ name, fileType, url, folderId, userID }) {
     const file = await prisma.file.create({
@@ -114,8 +85,6 @@ async function deleteFile(fileId) {
 }
 
 export {
-    getLoginUser,
-    postRegisterUser,
     insertFile,
     getFiles,
     getUserFiles,
