+++ src/db/queries/prisma.js
@@ -1,0 +1,4 @@
+import { PrismaClient } from '@prisma/client';
+const prisma = new PrismaClient();
+
+export default prisma;
